client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["bruce <cnhxunoo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = {extras = ["http2"], version = "^0.27.0"}
grpcio = "^1.63.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
flake8 = "^7.0.0"
yapf = "^0.40.2"
ruff = "^0.4.4"
grpcio-tools = "^1.63.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
